refactor: create cli pkg (https://github.com/Taiki130/ecsexec/issues/12)
