+++ pkg/cli/profile.go
@@ -1,0 +1,37 @@
+package cli
+
+import (
+	"strings"
+
+	"github.com/aws/aws-sdk-go-v2/config"
+	"github.com/manifoldco/promptui"
+	"gopkg.in/ini.v1"
+)
+
+func SelectProfile() (string, error) {
+	l := "Select Profile"
+
+	fname := config.DefaultSharedConfigFilename()
+	f, err := ini.Load(fname)
+	if err != nil {
+		return "", err
+	}
+
+	var profiles []string
+	for _, v := range f.Sections() {
+		if len(v.Keys()) != 0 {
+			profile := strings.Split(v.Name(), " ")[1]
+			profiles = append(profiles, profile)
+		}
+	}
+
+	prompt := promptui.Select{
+		Label: l,
+		Items: profiles,
+	}
+	_, result, err := prompt.Run()
+	if err != nil {
+		return "", err
+	}
+	return result, nil
+}
